Fixes for referenced image bug
previous commit had a bug where the files were all being created, but still referenced the testimage variable.  This made the naming conventions of the csvs created correct, but the data all from a single image.

This is now fixed by making each file loop pass its name as a string that is then set as the testimage in the loop.  works perfectly now.

Sim