more networking
So what we CAN do is add a text file to this repository, then make the program get it's own latest straight thru github.  If the textfile can update itself thru github then we can use github to force the program to get the latest dynamic link to use when downloading from google drive.